requirements.txt
discord==2.3.2
discord.py==2.6.4
python-dotenv==1.1.1
setproctitle==1.3.7
google-genai==1.45.0
